+++ js/templates/-attempts-ended.js
@@ -1,5 +1,6 @@
 import createElement from './../-create-elem';
 import renderScreen from "./../-render-screen";
+import welcome from "./-welcome";
 
 const attemptsEnded = createElement`<section class="main main--result">
     <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
@@ -12,5 +13,12 @@ const attemptsEnded = createElement`<section class="main main--result">
 const renderAttemptsEnded = () => {
   renderScreen(attemptsEnded);
 };
+
+document.addEventListener(`click`, (evt) => {
+  if (evt.target.classList.contains(`main-replay`)) {
+    evt.preventDefault();
+    welcome();
+  }
+});
 
 export default renderAttemptsEnded;
