+++ js/templates/-result-success.js
@@ -1,5 +1,6 @@
 import createElement from './../-create-elem';
 import renderScreen from "./../-render-screen";
+import welcome from "./-welcome";
 
 const resultSuccess = createElement`<section class="main main--result">
     <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
@@ -16,5 +17,13 @@ const resultSuccess = createElement`<section class="main main--result">
 const renderResultSucces = () => {
   renderScreen(resultSuccess);
 };
+
+
+document.addEventListener(`click`, (evt) => {
+  if (evt.target.classList.contains(`main-replay`)) {
+    evt.preventDefault();
+    welcome();
+  }
+});
 
 export default renderResultSucces;
