+++ js/templates/-time-out.js
@@ -1,5 +1,6 @@
 import createElement from './../-create-elem';
 import renderScreen from "./../-render-screen";
+import welcome from "./-welcome";
 
 const timeOut = createElement`<section class="main main--result">
     <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
@@ -13,5 +14,12 @@ const timeOut = createElement`<section class="main main--result">
 const renderTimeOut = () => {
   renderScreen(timeOut);
 };
+
+document.addEventListener(`click`, (evt) => {
+  if (evt.target.classList.contains(`main-replay`)) {
+    evt.preventDefault();
+    welcome();
+  }
+});
 
 export default renderTimeOut;
