+++ js/templates/-welcome.js
@@ -1,5 +1,6 @@
-import createElement from './../-create-elem';
+import createElement from "./../-create-elem";
 import renderScreen from "./../-render-screen";
+import artist from "./-artist-module";
 
 const welcomeElement = createElement`<section class="main main--welcome">
     <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
@@ -15,6 +16,12 @@ const welcomeElement = createElement`<section class="main main--welcome">
 const renderWelcomeScreen = () => {
   renderScreen(welcomeElement);
 };
+
+document.addEventListener(`click`, (evt) => {
+  if (evt.target.classList.contains(`main-play`)) {
+    artist();
+  }
+});
 
 export default renderWelcomeScreen;
